OughtaFocus: CheckStyling.  Split out BrentFocuser into its own file.
